Extract institution serialization to common project
