+++ common/src/main/resources/application.conf
@@ -1,3 +1,5 @@
+include "serialization.conf"
+
 db {
   driver = "slick.driver.PostgresDriver$"
