+++ common/src/main/resources/serialization.conf
@@ -1,0 +1,22 @@
+akka {
+  actor {
+    allow-java-serialization = off
+
+    serializers {
+      institution-commands = "hmda.serialization.institution.InstitutionCommandsSerializer"
+      institution-events = "hmda.serialization.institution.InstitutionEventsSerializer"
+    }
+    serialization-bindings {
+      "hmda.model.institution.Institution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$CreateInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$ModifyInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$GetInstitution" = institution-commands
+      "hmda.messages.institution.InstitutionCommands$DeleteInstitution" = institution-commands
+
+      "hmda.messages.institution.InstitutionEvents$InstitutionCreated" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionModified" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionDeleted" = institution-events
+      "hmda.messages.institution.InstitutionEvents$InstitutionNotExists" = institution-events
+    }
+  }
+}
