+++ hmda/src/main/resources/persistence.conf
@@ -1,3 +1,5 @@
+include "serialization.conf"
+
 akka {
 
   persistence {
@@ -10,26 +12,7 @@ akka {
     }
   }
 
-  actor {
-    allow-java-serialization = off
 
-    serializers {
-      institution-commands = "hmda.persistence.serialization.institution.InstitutionCommandsSerializer"
-      institution-events = "hmda.persistence.serialization.institution.InstitutionEventsSerializer"
-    }
-    serialization-bindings {
-      "hmda.model.institution.Institution" = institution-commands
-      "hmda.messages.institution.InstitutionCommands$CreateInstitution" = institution-commands
-      "hmda.messages.institution.InstitutionCommands$ModifyInstitution" = institution-commands
-      "hmda.messages.institution.InstitutionCommands$GetInstitution" = institution-commands
-      "hmda.messages.institution.InstitutionCommands$DeleteInstitution" = institution-commands
-
-      "hmda.messages.institution.InstitutionEvents$InstitutionCreated" = institution-events
-      "hmda.messages.institution.InstitutionEvents$InstitutionModified" = institution-events
-      "hmda.messages.institution.InstitutionEvents$InstitutionDeleted" = institution-events
-      "hmda.messages.institution.InstitutionEvents$InstitutionNotExists" = institution-events
-    }
-  }
 }
 
 cassandra-journal {
