+++ hmda/src/test/scala/hmda/model/filing/ts/TsGenerators.scala
@@ -2,7 +2,7 @@ package hmda.model.filing.ts
 
 import org.scalacheck.Gen
 import hmda.generators.CommonGenerators._
-import hmda.model.filing.FilingGenerators._
+import hmda.model.institution.InstitutionGenerators._
 
 object TsGenerators {
